#44 front 페이지 추가 - wefluxsecurity로 변경
