+++ authentication/src/main/resources/application.yml
@@ -4,9 +4,7 @@ logging:
   level:
     root: info
     org:
-      springframework:
-        jdbc:
-          core: debug
+      springframework: trace
   pattern:
     console: "%d{HH:mm:ss} [%t][%-5level] %msg \\(%F:%L\\)%n"
 spring:
