+++ front/src/main/resources/application.yml
@@ -6,6 +6,8 @@ server:
 logging:
   level:
     root: info
+    org:
+      springframework: trace
 spring:
   security:
     oauth2:
